third_party/blink/web_tests/fast/events/mouse-event-source-device-event-sender-expected.txt

Test that mouse events from real mouse device should have a sourceCapabilities and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

mousemove
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
mousedown
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
mouseup
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
click
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
mousedown
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
mouseup
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
click
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
dblclick
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
